requirements.txt
langchain
openai
python-dotenv
pyyaml
streamlit
langsmith